chore: resolve circular imports (#410)
