+++ projects/ng-aquila/src/dropdown/dropdown.control.ts
@@ -1,4 +1,5 @@
 import { NxFormfieldControl } from '@aposin/ng-aquila/formfield';
+import { Observable, Subject } from 'rxjs';
 
 export abstract class NxDropdownControl extends NxFormfieldControl<any> {
   /**
@@ -7,7 +8,14 @@ export abstract class NxDropdownControl extends NxFormfieldControl<any> {
    * Note: Please make sure the value you bind is an array. If not an error is thrown! */
   isMultiSelect: boolean = false;
 
+  readonly filterChanges!: Subject<any>;
+
+  readonly _closedStream!: Observable<void>;
+
+  abstract filterFn(search: string, itemValue: string): boolean;
+
+  abstract valueFormatter(value: any): string;
+
   /** @docs-private */
   abstract formatValue?(value: any): string;
-
 }
