+++ projects/ng-aquila/src/progress-stepper/progress-stepper.models.ts
@@ -1,0 +1,2 @@
+/** Direction of the multi stepper */
+export type NxMultiStepperDirection = 'vertical' | 'horizontal';
