+++ projects/ng-aquila/src/tabs/tab-group-base.ts
@@ -1,6 +1,6 @@
 import { Directive } from '@angular/core';
 import { Subject } from 'rxjs';
-import { NxTabsAppearance } from './tab-group';
+import { NxTabsAppearance } from './tabs.models';
 
 /** @docs-private */
 @Directive()
